refactor(discord): retire internal abort signal helper
