+++ extensions/discord/runtime-api.monitor.ts
@@ -48,7 +48,6 @@ export {
   DISCORD_DEFAULT_INBOUND_WORKER_TIMEOUT_MS,
   DISCORD_DEFAULT_LISTENER_TIMEOUT_MS,
   isAbortError,
-  mergeAbortSignals,
   normalizeDiscordInboundWorkerTimeoutMs,
   normalizeDiscordListenerTimeoutMs,
   runDiscordTaskWithTimeout,
